checkpoint; working demo
The demo brings up a controller with 2 outs and 1 in, then periodically
sets the levels of the outputs to random values. The processor picks
these up and the effect can be seen in a JACK oscilloscope program.

Next step is to implement something that would require deallocation by
the controller thread, i.e. one of
- Removing an input or an output
- Creating and removing loopback streams