change to language, reorg
The language is now entirely f32-valued. The former "now" term which
would give a u64 of the current sample was removed in favour of simply
offering various oscillators parameterized by frequency and phase. Now
to get a pulse every second, for instance, you could use a sawtooth of
frequency 1, then multiply and add to it appropriately. That's much
nicer than taking the current sample modulo the sample rate (the
programmer shouldn't need to deal with the sample rate so much.. maybe
only in the context of loading up PCM data).